Reword bullet item 9 to emphasize the use of the dropdown menu and not the filter search box.
The current bullet item (number 9) does not make it clear that the dropdown functionality should be used .  By mistake, I was trying to use the filter search box instead, which does not produce the expected results when "AWS managed - job function" is entered into it. (The filter search box is buggy and displays incorrect results.)  I feel it will be helpful to stress that there is a drowdown with a check box that needs to be used. Thanks.