Update id_root-user.md
Grammatical changes, merely to make the content clearer.

I assumed the docs were wrong in claiming that deactivated keys do not count towards a limit, but upon looking to fix them I realised they were correct, however when skimmed over the grammar makes it easy to assume they do not count. The use of 'and' instead of a negative makes it sound like the effects of deactivating have changed their effect on the key limit.

I would have ignored this if re-reading the second time I had noticed the true meaning. However, others may also be skimming over and making the same assumption as me.