Update access_policies_access-advisor-view-data.md
I noted that this URL and others point to AWS docs with a message stating: ```AWS CLI version 2, the latest major version of AWS CLI, is now stable and recommended for general use. To view this page for the AWS CLI version 2, click here. For more information see the AWS CLI version 2 installation instructions and migration guide.```

if this is true then the URL should probably point to the new docs, and the other URLs in this doc should be updated. I did note though that the original URL points to `/latest`, so maybe this should be updated elsewhere in AWS docs and this PR closed instead?