Fix multi-value context key DynamoDB example
The DynamoDB multi-value context key examples omit the fact that
`GetItem`, `PutItem`, `UpdateItem` (and likely other operations) always include
the primary key attributes in the `dynamodb:Attributes` context key. Therefore:

1. The `ForAllValues` example provides a false sense of security since it does
   not protect for the case when the user implicitly requests all attributes by
   not specifying a `ProjectionExpression`.
2. The `ForAnyValue` example provides a condition that will never evalute to
   false since it includes the partition key in the value list.

This commit fixes the previously mentioned problems in the examples.
Additionally, the `ForAnyValue` example is updated to use the `UpdateItem`
operation instead of `PutItem` since it has more semantic meaning.

Other minor changes like the order of appearance of `Warning` and `Important`
messages and the reference to the DynamoDB documentation are updated to more
logical places.